Moved all renderers to Patches
