--- Lunity/Client/Renderer/Renderer.cpp
@@ -1,5 +1,0 @@
-//
-// Created by lukax on 10.07.2021.
-//
-
-#include "Renderer.h"
